cloud_run_runner/requirements.txt
pandas
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-aiplatform
